Fixed parse error in pm_write.php
Added metarefresh option to html_draw_top(). Call it with the option refresh=<seconds> in the call to the function the same you would for target frame etc.

Added notification icon to nav.php so it will display an icon (pmnew.png) next to the link to the PM inbox if you have any new PMs. Nav,php takes advantage of html_draw_top()'s refresh option to refresh itself once every 60 seconds. Not sure about this, but it can easily be removed if it doesn't go down well.