Changed the (closed) text to a padlock icon. Better icon coming soon. Once someone draws one.
Added an additional field to the PM table that logs if the user has seen a popup notification for new PMs. The old method of PM_NEW constant is only handled by the thread list now and is used to display the [x unread] next to the PM Inbox link.

Note: New upgrade-04-to-041.sql file needs to be run against the DB.