Fixed bug in bh_session_check which was inserting the LAST_LOGON twice and causing a duplicate index error.
Made double sure that only the Queen can currently delete and create forums. This may change in order to allow users to sign up and create their own forums, but that isn't the line I want to go down at the moment and Beehive is good enough to allow people to make use of the files to create their own signup / subscription page if they want to.