Fixed create_forum() was not up to date and wasn't creating the new ALLOW_EMAIL and ALLOW_PM fields in USER_PREFS table.
Fixed admin_menu.php not passing webtag to admin_foruims.php and admin_forums.php not reading the webtag. It isn't needed by the page as it displays server specific data and not forum specific but still we use it to make sure the page is displayed in the user's correct style etc.

Fixed nav.php not loading forum settings correctly and therefore showing in the wrong style compared to what the user has picked.