"Fixed" The Unread Message count on My Forums. Ha!
Added function delete_attachment_by_aid which fetches all attachments with the same AID (attached to the same message in theory) and then deletes them all.

Changed delete_attachment() so it gets the UID of the user from the session rather than from the function parameters

Fixed bug in PMs when deleting a PM that contains an attachment.