Fixed auto-nesting problem, where:
<ol>
  <li>1
  <li>2
</ol>

would become:

<ol>
  <li>1
  <ul><li>2</li></ul></li>
</ol>

Also fixed stripslashes bug in <code> - \" was being converted to ".

Also defined variables for the code/quote text at the top of the document, so they can be set to use the forum's default language or something. At the moment they're just defined to be "code:" and "quote:", though, as before.