Changed the error handler to use PHP's own defined error types for now.
Updated install and upgrade script so that they don't fetch the mysql error string because we do that in install.php as well

Included error handler in install.php to catch SQL errors which cause the install to fail. Not sure how productive this is going to be, so might remove it later.