Removed some of the settings from both admin_default_forum_settings.php and admin_forum_settings.php. Only duplicate settings now are the Guest account and spidering options.
Change save_forum_settings() and save_default_forum_settings() functions so they remove all the old entries from the database before saving the new ones.

Fixed bug in gzip encode function that was checking for database entry rather than reading from config.inc.php to check if gzip should be enabled or not.

Changed Admin pages "Default Forum Settings" to "Global Forum Settings" and added a warning text to the bottom of admin_default_forum_settings.php (above the Save button) stating what the settings do.

Fixed bug in session.inc.php which meant that the logon form wouldn't be drawn if the user session expired on certain pages.

Added a function to check if the user cookie bh_remember_username is set so we can tell if the user has visited before or not before we decide to try and log them in as a guest or not.

Update new-install.php and upgrade-04-to-05.php to make the settings stick for the right global / per-forum level.