Added donation link to the Beehive bar.
Added entry for Post Approval to be shown in the admin log

Uber updated thumbnail creation function that checks for the GD version and falls back to older functions if they don't exist. Also checks correctly with php_info() to see if the required image support is enabled.

Updated all the style sheets to contain the same classes and the correct colours for each theme.

Fixed thread list showing the wrong borders around the folders.