New search!
Completely scrapped the old method of doing a LIKE based search directly on POST_CONTENT as it is inefficient and drags MySQL to it's knees with a relatively large database (Teh Forum's 500MB craps itself when anyone searches) so instead we now have a new table which has indexed keywords of the posts that we match against. This has meant the Search form is much more reduced in functionality than previously, but it's quicker and less resource hungry which is what matters.