Worming was broken seemingly and checking through the CVS logs it seems it was broken before the release of 0.5? Me = confused.
Added a quick dirty function to server.inc.php to fetch the current time for use in measuring page execution time.

Trying to solve incredibly slow loading times for admin_user.php, but no idea what is causing it as the SQL queries all run really quickly, but the page sits there for ages still. Again, me = confused.

Changed admin_folders so it fetches the thread count seperatly in an effort to speed up display of the page.