Expanded DEDUPE table into USER_TRACK which tracks the last time the user posted and searched in addition to the de-duplication code.
Added functionality to restrict how often a user can post or search.

Added some javascript to disable the Find button once a user clicks it. This could be extended to other forms if we need it, but not right now. CSS for this is only in the default style for the moment. I'll update the other styles later some time.

Fixed text input fields not working in admin_forum_settings.php

Fixed undefined variable error when trying to save global forum settings.

Changed search results page header to show "Error" when there is an error, rather than still show "Search results"