Added two options to the installer to force deletion of the the tables that Beehive uses during a new install and to skip creation of the dictionary.
Added check to dictionary class so it can inform the user if the dictionary is not setup correctly.

Changed db connection scripts to only trigger errors when there is no @ prepended to the calling function name. Used by the installation scripts to prevent Beehive from triggering an error display when encountering an error during installation.