More changes so that emoticons are parsed per page-view rather than on post submission.
A lot of the emoticon code is now in a class of it's own. I still have to remove the emoticon adding/removing code from fixhtml as well as code in some new allowance for things like <noemots> and no-emoticons-in-<code>-and-<spoiler>-and-I-don't-know-what-else-tags.

oh, and fixed a small bug in display_emoticons.php