Fixed undefined variable $db_post_delete error. Should have been $db_post_update
Fixed wrong variable name in messages_get() causing not everything to be set correctly, which was causing errors in the search results.

Changed search to show the number of matches we found in the database rather than a maximum of 50 regardless of whatever were found.

Reformmated search.php's checking for POST and GET variables.