Fixed forum_create was creating a visitor_log table which is unneeded.
Fixed anon logon not working correctly. The query to NULL the LAST_LOGON column for the user was not actually being called at all.

Fixed bug #20050414 where the timestamp for the last visit was incorrect.

Fixed the illusion of day light savings appearing to be enabled when it wasn't in forum_options.php