Updated preview of admin_make_style.php to reflect more like 0.6 looks. Not added the stats display though.
Rewrote most of the inards of admin_make_style.php so it has some error pages to fall back on with instructions on downloading the created style.css rather than forcing it on the user.

Language file updates

Added function to make_style.inc.php that actually does the style saving.