Fixed USER_PERM_PILLORIED permission always showing edit button even when the user doesn't have permission to edit the post.
Added function to fetch all folders in a drop down and ignoring permissions and allowed types and made admin_rss_feeds.php use it. Useful because you might want to add a feed with a logon which isn't the same as your own so you won't know about it's user permissions.

Updated x-hacker language file.