Fixed parse error in llogon.php. Oops!
Removed upgrade-05pr1-to-05.php because I'm not going to support the use of 0.5PR1 anymore. If you're still using it UPGRADE ALREADY!!

Fixed error in ewiki integration script where the query was looking for the old tables. It now looks for the 0.6 named tables and allows you to edit the $forum_fid being used if you want to link your ewiki install with a specific webtag.

Adding some browser work arounds and fixes to the installer. Hoping to combat the blank screen timeout error caused by the dictionary taking a long time to process.

Added CLI install method to 0.4-to-0.5 install script incase anyone ever wants to use it.