Updated deprecated string finding script which is less resource intensive.
Fixed bug #1196690 where it was impossible to navigate beyond page 2 of the links pages.
Fixed bug #1196367 missing index makesticky in language files. Must have deleted it accidentally.
Fixed bug #1196373 thread list was not translating with the rss feed not using the user sess correctly.
Fixed bug #1196686 made all the links.php actions redirect to itself so that the links are reloaded.