Changed rss_add_feed() to set the last run date for a new feed to be Beehive 0.1's release date so that it is forced to run straight away when the thread list is next reloaded by someone.
Fixed new-install.php not working correctly when run from a browser. Changed upgrade--04-to-05.php to use the better dictionary import code with the correct path detection. Changed all installer scripts so they require the CLI install mode to be run from within the install folder. This is so we can correctly try and do the LOAD DATA query for the dictionary.

Made new-install.php and install.php independently set the admin_logon and admin_nickname variables with the logon being uppercased automatically and the nickname left in the case the user specifies.