Removed directory creation gubbins for text captcha and attachments from admin_default_forum_settings.php and added a function that is called by the script and all the other attachment scripts to check and create the directory. The text captcha class already does this.
Fixed error handler falling over itself if the $_POST variables contains an array. Handled by a function in form.inc.php that can be used by other scripts if required.

Fixed error in get_request_uri which trips over $_GET variable containing an array. Handled by a seperate function to the one in form.inc.php as this one is required to construct a URL query from them.

Removed erranous space from end of constants.inc.php that trips up if the output buffering is disabled.

Fixed post content preview being chopped incorrectly in search results and fixed search not working as expected when a user searches for words by a user and the word is in the MySQL stop list.