Simplified the calls to the email notification functions requiring a msg and splitting it into tid.pid when the function is called by joining the tid & pid!
Fixed attachments being deleted prevents moderators from gaining access to the edit attachments page of a post for another user.

Changed the RSS feed so it duplicates the headline in the quote tag so if it gets truncated by Beehive's 64 char limit on thread titles we still get to see it all.